separate: func
